Fuse futures before loop in process_events_async
As per the futures::select documentation:

> Futures and streams which are not already fused can be fused using the
> .fuse() method. Note, though, that fusing a future or stream directly
> in the call to select! will not be enough to prevent it from being
> polled after completion if the select! call is in a loop, so when
> select!ing in a loop, users should take care to fuse() outside of the
> loop.

In our case, this isn't strictly required as we would always have a new
instance of the futures on each poll, but just to be safe, we follow the
suggestion anyway. As a result, we generate a new future for each one
that has completed, so that we can continue with execution.